Add external timed text source using a passed file descriptor
o Also fixed a minor issue where the file length should be of type off64_t rather than size_t

o related-to-bug: 5542712

+++ media/libstagefright/AwesomePlayer.cpp
@@ -2292,7 +2292,7 @@ status_t AwesomePlayer::invoke(const Parcel &request, Parcel *reply) {
             }
             int fd         = request.readFileDescriptor();
             off64_t offset = request.readInt64();
-            size_t length  = request.readInt64();
+            off64_t length  = request.readInt64();
             String8 mimeType(request.readString16());
             return mTextDriver->addOutOfBandTextSource(
                     fd, offset, length, mimeType);
